refactor config and proxy-server into separate files
Proxy-server needs to be started before malware-scanner
