+++ cloudrun-malware-scanner/bootstrap.sh
@@ -47,13 +47,6 @@ apt-get -qqy install --no-install-recommends clamav-daemon clamav-freshclam
 export PATH="$PATH:$HOME/.local/bin" # add pipx locations to path.
 pipx install cvdupdate
 
-# The node server includes a reverse proxy which adds authentication
-# to requests to GCS REST API, allowing freshclam to access the GCS
-# CVD mirror as if it was an unauthenticated HTPP server
-#
-export PROXY_PORT=8888
-PROXY_SERVER_ADDRESS=127.0.0.1:${PROXY_PORT}
-
 # Ensure clamav services are shut down, as we do not have the config files set up yet.
 service clamav-daemon stop &
 service clamav-freshclam stop &
@@ -81,6 +74,15 @@ if ! gsutil ls "gs://${CVD_MIRROR_BUCKET}/" > /dev/null ; then
     Log ERROR main "ClamCvdMirrorBucket=${CVD_MIRROR_BUCKET} does not exist or is not readable"
     exit 1
 fi
+
+# Start the reverse proxy which adds authentication
+# to requests to GCS REST API, allowing freshclam to access the GCS
+# CVD mirror bucket as if it was an unauthenticated HTPP server
+#
+export PROXY_PORT=8888
+PROXY_SERVER_ADDRESS=127.0.0.1:${PROXY_PORT}
+npm start-proxy "${CONFIG_FILE}" &
+sleep 5
 
 # This function is used to update clam and freshclam config files.
 # Use by specifying the config file on the command line and
