+++ cloudrun-malware-scanner/package.json
@@ -1,10 +1,11 @@
 {
   "name": "gcs-malware-scanner",
-  "version": "2.2.0",
+  "version": "2.3.0",
   "description": "Service to scan GCS documents for the malware and move the analyzed documents to appropriate buckets",
   "main": "index.js",
   "scripts": {
     "start": "node server.js",
+    "start-proxy": "node gcs-proxy-server.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "eslint": "eslint *.js",
     "eslint-fix": "eslint --fix *.js"
