+++ cloudrun-malware-scanner/server.js
@@ -24,7 +24,8 @@ const metrics = require('./metrics.js');
 const util = require('node:util');
 const execFile = util.promisify(require('node:child_process').execFile);
 const {setTimeout} = require('timers/promises');
-const httpProxy = require('http-proxy');
+// eslint-disable-next-line no-unused-vars
+const {Config, readAndVerifyConfig} = require('./config.js');
 
 const PORT = process.env.PORT || 8080;
 const CLAMD_HOST = '127.0.0.1';
@@ -50,15 +51,7 @@ const MAX_FILE_SIZE = 500000000; // 500MiB
  * Values are read from the JSON configuration file.
  * See {@link readAndVerifyConfig}.
  *
- * @type {{
- *    buckets: Array<
- *      {
- *        unscanned: string,
- *        clean: string,
- *        quarantined: string
- *       }>,
- *    ClamCvdMirrorBucket: string
- *  }}
+ * @type {Config}
  */
 const BUCKET_CONFIG = {
   buckets: [],
@@ -71,10 +64,6 @@ app.use(express.json());
 const scanner = clamd.createScanner(CLAMD_HOST, CLAMD_PORT);
 const storage = new Storage({userAgent: `${pkgJson.name}/${pkgJson.version}`});
 const googleAuth = new GoogleAuth();
-
-// access token for GCS requests - will be refreshed every 50 mins
-let accessToken;
-const ACCESS_TOKEN_REFRESH_INTERVAL = 50*60*1000;
 
 /**
  * Route that is invoked by Cloud Run when a malware scan is requested
@@ -324,95 +313,6 @@ async function moveProcessedFile(filename, isClean, config) {
 }
 
 /**
- * Read configuration from JSON configuration file.
- * and store in BUCKET_CONFIG global
- *
- * @async
- * @param {string} configFile
- */
-async function readAndVerifyConfig(configFile) {
-  logger.info(`Using configuration file: ${configFile}`);
-
-  try {
-    const config = require(configFile);
-    delete config.comments;
-    Object.assign(BUCKET_CONFIG, config);
-  } catch (e) {
-    logger.fatal(
-        {err: e},
-        `Unable to read JSON file from ${configFile}`);
-    throw new Error(`Invalid configuration ${configFile}`);
-  }
-
-  if (BUCKET_CONFIG.buckets.length === 0) {
-    logger.fatal(`No buckets configured for scanning in ${configFile}`);
-    throw new Error('No buckets configured');
-  }
-
-  logger.info('BUCKET_CONFIG: '+JSON.stringify(BUCKET_CONFIG, null, 2));
-
-  // Check buckets are specified and exist.
-  let success = true;
-  for (let x = 0; x < BUCKET_CONFIG.buckets.length; x++) {
-    const config = BUCKET_CONFIG.buckets[x];
-    for (const bucketType of ['unscanned', 'clean', 'quarantined']) {
-      if ( !(await checkBucketExists(
-          config[bucketType],
-          `config.buckets[${x}].${bucketType}`))) {
-        success=false;
-      }
-    }
-    if (config.unscanned === config.clean ||
-        config.unscanned === config.quarantined ||
-        config.clean === config.quarantined) {
-      logger.fatal(
-          `Error in ${configFile} buckets[${x}]: bucket names are not unique`);
-      success = false;
-    }
-  }
-  if ( !(await checkBucketExists(
-      BUCKET_CONFIG.ClamCvdMirrorBucket,
-      'ClamCvdMirrorBucket'))) {
-    success=false;
-  }
-
-  if (!success) {
-    throw new Error('Invalid configuration');
-  }
-}
-
-
-/**
- * Check that given bucket exists. Returns true on success
- *
- * @param {string} bucketName
- * @param {string} configName
- * @return {Promise<boolean>}
- */
-async function checkBucketExists(bucketName, configName) {
-  if (!bucketName) {
-    logger.fatal(`Error in config: no "${configName}" bucket defined`);
-    success = false;
-  }
-  // Check for bucket existence by listing files in bucket, will throw
-  // an exception if the bucket is not readable.
-  // This is used in place of Bucket.exists() to avoid the need for
-  // Project/viewer permission.
-  try {
-    await storage.bucket(bucketName).getFiles(
-        {maxResults: 1, prefix: 'zzz', autoPaginate: false});
-    return true;
-  } catch (e) {
-    logger.fatal(
-        `Error in config: cannot view files in "${
-          configName}" : ${bucketName} : ${e.message}`);
-    logger.debug({err: e});
-    return false;
-  }
-}
-
-
-/**
  *  Wait up to 5 mins for ClamD to respond
  */
 async function waitForClamD() {
@@ -435,64 +335,6 @@ async function waitForClamD() {
 }
 
 /**
- * Set up a reverse proxy to add authentication to HTTP requests from
- * freshclam and proxy it to the GCS API
- */
-async function setupGcsReverseProxy() {
-  // Get an OAuth2 access token and refresh it every 50mins.
-  accessToken = await googleAuth.getAccessToken();
-  
-  setInterval(async () => {
-    logger.info(`Refreshing Oauth2 Access Token for GCS proxy.`);
-    accessToken = await googleAuth.getAccessToken();
-  }, ACCESS_TOKEN_REFRESH_INTERVAL);
-
-  const proxy = httpProxy.createProxyServer({
-    target: 'https://storage.googleapis.com/',
-    changeOrigin: true,
-    autoRewrite: true,
-    secure: true,
-    ws: false,
-  });
-
-  // Error handling...
-  proxy.on('error', function(err, req, res) {
-    let statusCode = 500;
-    if (res && res.statusCode && res.statusCode != 200) {
-      statusCode = res.statusCode;
-    }
-    logger.error(`Failed to proxy to GCS for path ${req.url}, returning code ${
-      statusCode}:  ${err}`);
-    res.writeHead(statusCode, {
-      'Content-Type': 'text/plain',
-    });
-    res.end(`Failed to proxy to GCS: status ${statusCode}\n`);
-  });
-
-  // Add auth header/
-  proxy.on('proxyReq', function(proxyReq, req, res) {
-    if (proxyReq.path.startsWith(
-        '/' + BUCKET_CONFIG.ClamCvdMirrorBucket + '/')) {
-      logger.info(`Proxying request for ${proxyReq.path} to GCS`);
-      proxyReq.setHeader('Authorization', 'Bearer ' + accessToken);
-    } else {
-      logger.error(
-          `Denying Proxy request for ${proxyReq.path} to GCS - invalid path`);
-      res.writeHead(404, {
-        'Content-Type': 'text/plain',
-      });
-      res.end('Failed to proxy to GCS - invalid path: status 404\n');
-    }
-  });
-
-  const PROXY_PORT = process.env.PROXY_PORT || 8888;
-
-  proxy.listen(PROXY_PORT, 'localhost');
-  logger.info(
-      `GCS authenticating reverse proxy listenting on port ${PROXY_PORT}`);
-}
-
-/**
  * Perform async setup and start the app.
  *
  * @async
@@ -511,9 +353,9 @@ async function run() {
   } else {
     configFile = './config.json';
   }
-  await readAndVerifyConfig(configFile);
+  const config = await readAndVerifyConfig(configFile);
 
-  await setupGcsReverseProxy();
+  Object.assign(BUCKET_CONFIG, config);
 
   await waitForClamD();
